+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 * Fix the issue where the tile URL cannot be correctly parsed with the HTTPS protocol when using an nginx proxy service (https://github.com/maptiler/tileserver-gl/pull/1578) by @dakanggo
 * Use jemalloc as memory allocator in the docker image (https://github.com/maptiler/tileserver-gl/pull/1574) by @MichielMortier
 * Rasters: Add tileSize to TileJSON (https://github.com/maptiler/tileserver-gl/pull/1559) by @roblabs
-* Allow a 'sparse' option per data source (https://github.com/maptiler/tileserver-gl/pull/1558( by @acalcutt
-* Updates Maplibre-gl-js to v5.6.0 and adds color-relief support (note: this is not yet supported by maplibre-native). (https://github.com/maptiler/tileserver-gl/pull/1591)
+* Allow a 'sparse' option per data source (https://github.com/maptiler/tileserver-gl/pull/1558) by @acalcutt
+* Updates Maplibre-gl-js to v5.6.0 and adds color-relief support (note: this is not yet supported by maplibre-native) (https://github.com/maptiler/tileserver-gl/pull/1591)
 
 ## 5.3.0
 * Fix - Include public\resources js files on npm publish by specifying included files in package.json (https://github.com/maptiler/tileserver-gl/pull/1490) by @acalcutt