+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.5.0-pre.2
 * Add S3 support for PMTiles with multiple AWS credential profiles (https://github.com/maptiler/tileserver-gl/pull/1779) by @acalcutt
 * Create .aws directory passthrough folder in Dockerfile (https://github.com/maptiler/tileserver-gl/pull/1784) by @acalcutt
+* Update eslint to v9 (https://github.com/maptiler/tileserver-gl/pull/1473) by @acalcutt
 * Fix Renderer Crashes from Failed Fetches (https://github.com/maptiler/tileserver-gl/pull/1798) by @acalcutt
 
 ## 5.4.0